phoenix_crm/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-jose[cryptography]
python-dotenv
pydantic[email]
email-validator